refactor: production-grade architecture overhaul v0.2.0 → v0.3.0
BREAKING CHANGE: Schema migration from REAL to TEXT for exact Decimal arithmetic.
All money columns now store Decimal strings to eliminate float rounding.

Core changes:
- AppContext DI refactor: eliminate global mutable singletons (get_store,
  reset_store, telemetry singletons). All dependencies injected via ctx.obj.
- Schema migration runner: numbered .sql files in migrations/ with
  schema_version tracking. Unblocks future schema changes safely.
- Config dir permissions: enforce 0o700 on ~/.ai-spend.
- Provider retry logic: exponential backoff _make_request with 429
  Retry-After respect, fail-fast on 4xx, retry on 5xx/connect errors.

CLI additions:
- sync --dry-run / -n: preview records without writing
- sync --provider / -p: sync a single provider
- config validate: test API credentials without syncing
- import: round-trip JSON/CSV import with atomic transaction
- stats --json: structured telemetry output

Observability:
- Structured JSON logging (ai_spend.log) with configurable verbosity
- Sync operation logs: sync_start, sync_success, sync_failure
- Provider retry logs: request_retry, request_failed

Quality gates:
- mypy strict mode added to CI (type-check job in ci.yml)
- ruff enforced across src/ and tests/
- Coverage: 96.6% (311 tests, threshold 90%)

New files:
- src/ai_spend/context.py (AppContext dataclass)
- src/ai_spend/log.py (JsonFormatter, configure_logging)
- src/ai_spend/migrations/001_initial.sql, 002_decimal_money.sql
- tests/test_log.py, test_main.py, test_migrations.py

+++ src/ai_spend/providers/manual.py
@@ -3,6 +3,7 @@
 from __future__ import annotations
 
 from datetime import date
+from decimal import Decimal
 
 from ai_spend.models import ProviderType, UsageRecord
 from ai_spend.providers.base import BaseProvider
@@ -27,7 +28,7 @@ class ManualProvider(BaseProvider):
     def create_entry(
         self,
         model: str,
-        cost_usd: float,
+        cost_usd: Decimal,
         entry_date: date | None = None,
         note: str = "",
     ) -> UsageRecord:
